Integrate notmuch with Gnus
Some basic integration of notmuch with Gnus. The first thing ot note
is that notmuch is really set up for Emacs23, and using it with the
MacOS X version of GNU Emacs (22.1) is not possible. So we are stuck
with using this stuff with Aquamacs or some such. So the notmuch
integration is also enabled only when we are on Emacs23.

Even then, it is not working fully - C-c C-c is not working. Need to
look into it.